test(e2e): extract shared filter-LWW fixture helpers (#215)
